teste_models.py
from openai import OpenAI
import os
from dotenv import load_dotenv

load_dotenv()  # 👈 ISSO FALTAVA

client = OpenAI(
    api_key=os.getenv("GROQ_API_KEY"),
    base_url="https://api.groq.com/openai/v1"
)

models = client.models.list()

for m in models.data:
    print(m.id)